deletes createPulse and animatePulse func from the POIInfoViewController. swift file and adds it to the POIInfoViewController+PulseAnimation.swift file
