+++ quavi/quavi/Controllers/POI Info/POIInfoViewController.swift
@@ -114,36 +114,6 @@ class POIInfoViewController: UIViewController {
             }
         }
     }
-    
-    func createPulse(){
-           let position = easterEggButton.frame.size.width / 2
-           let circularPath = UIBezierPath(arcCenter: .zero, radius: 16, startAngle: 0, endAngle: 2 * CGFloat.pi, clockwise: true)
-           shapeLayer.path = circularPath.cgPath
-           shapeLayer.strokeColor = #colorLiteral(red: 0.7416164279, green: 0.5822635889, blue: 0.9156076312, alpha: 1)
-           shapeLayer.lineWidth = 30.0
-           shapeLayer.lineCap = .round
-           shapeLayer.position = CGPoint(x: position, y: position)
-           easterEggButton.layer.addSublayer(shapeLayer)
-           animatePulse(layer: shapeLayer)
-       }
-       
-    func animatePulse(layer:CAShapeLayer){
-           let scaleAnimation = CABasicAnimation(keyPath: "transform.scale")
-           scaleAnimation.duration = 2.0
-           scaleAnimation.fromValue = 0
-           scaleAnimation.toValue = 1
-           scaleAnimation.timingFunction = CAMediaTimingFunction(name: .easeInEaseOut)
-           scaleAnimation.repeatCount = .greatestFiniteMagnitude
-           layer.add(scaleAnimation, forKey: "scale")
-           
-           let opacityAnimation = CABasicAnimation(keyPath: #keyPath(CALayer.opacity))
-           opacityAnimation.duration = 2.0
-           opacityAnimation.fromValue = 1
-        opacityAnimation.toValue = 0
-           opacityAnimation.timingFunction = CAMediaTimingFunction(name: .easeOut)
-           opacityAnimation.repeatCount = .greatestFiniteMagnitude
-           layer.add(opacityAnimation, forKey: "opacity")
-       }
 }
 
 extension POIInfoViewController: UIScrollViewDelegate{
